// apps/proyecto-integrado/src/app/state/actions/wishlists/wishlists-action.types.enum.ts
export enum WishlistsActionTypes {
  CREATE_WISHLIST_REQUEST = '[Wishlists] Create Wishlist Request',
  CREATE_WISHLIST_SUCCESS = '[Wishlists] Create Wishlist Success',
  CREATE_WISHLIST_FAILURE = '[Wishlists] Create Wishlist Failure',

  GET_WISHLISTS_REQUEST = '[Wishlists] Get Wishlists Request',
  GET_WISHLISTS_SUCCESS = '[Wishlists] Get Wishlists Success',
  GET_WISHLISTS_FAILURE = '[Wishlists] Get Wishlists Failure',

  DELETE_WISHLIST_REQUEST = '[Wishlists] Delete Wishlist Request',
  DELETE_WISHLIST_SUCCESS = '[Wishlists] Delete Wishlist Success',
  DELETE_WISHLIST_FAILURE = '[Wishlists] Delete Wishlist Failure',

  UPDATE_WISHLIST_REQUEST = '[Wishlists] Update Wishlist Request',
  UPDATE_WISHLIST_SUCCESS = '[Wishlists] Update Wishlist Success',
  UPDATE_WISHLIST_FAILURE = '[Wishlists] Update Wishlist Failure',
}